audio visualization/requirements.txt
streamlit
sounddevice
scipy
stability-sdk
Pillow
librosa